refactor(symbols): move get_symbols_overview into pipeline (arch PR D)
PR-D of the SymbolQuery pipeline series. `get_symbols_overview`
joins `get_ranked_context` (PR-B) + `find_symbol` (PR-C) behind
`SymbolQueryPipeline`. All three symbol-shape MCP tools now dispatch
through a single seam — `handlers::*` entries are 3-line thin
wrappers around the pipeline.

What moved into `symbol_query/symbols_overview.rs`:
  - `run_symbols_overview`: the 5-stage orchestration body
    (argument parsing → structural fetch → budget guard →
    degraded-result classification (#183/#184) → response
    annotations).
  - `resolve_path_argument` + `insert_response_annotations`:
    the path-alias resolver and top-level annotation inserter,
    both used only here.

`handlers::get_symbols_overview` is now:

    pub fn get_symbols_overview(state: &AppState, arguments: &Value) -> ToolResult {
        SymbolQueryPipeline::new(state)
            .run(SymbolQueryRequest::SymbolsOverview { arguments })
    }

Deletion test: removing `symbol_query/symbols_overview.rs` would
re-disperse the 100-line orchestration + two private helpers back
into handlers.rs. The symbols-overview semantics concentrate here.

handlers.rs evolution:
  - Pre PR-B: 1575 LOC (god module)
  - Post PR-C: 707 LOC (find_symbol body removed)
  - Post PR-D: 570 LOC (get_symbols_overview body removed)

Pipeline now owns all three primary symbol-shape tools. PR-E
will tighten the temporary `pub(crate)` visibility on
`symbols::handlers` + `symbols::formatter::compact_symbol_bodies`
that PR-B/C opened, and document `SymbolQueryPipeline` in
CLAUDE.md / docs/architecture.md.

Build matrix:
  - cargo check --workspace --features scip-backend,semantic ✓ (1.19s)
  - cargo check --workspace --no-default-features ✓
  - cargo clippy --features semantic -- -D warnings ✓
  - cargo clippy --no-default-features -- -D warnings ✓
  - cargo fmt --all -- --check ✓
  - cargo test -p codelens-mcp --bin codelens-mcp --features semantic
    → 634 passed / 1 ignored / 0 failed (61.16s)
  - 4 get_symbols_overview integration tests (degraded_reason,
    file_not_indexed, indexed_no_symbols, unsupported_extension)
    all pass.

Net diff: +178/-141 LOC across 3 files (1 new).
